Update lexemesTypes.cpp
Got integers working, I wrote it to differentiate between hex,binary, decimal, octal digits. I do not know why we have to differentiate between the types but I got it to work nonetheless.As it stands this section works but might be able to be better written for coding style.